has to link to files in the website directory.
This is a shame, because when we develop the actual UI the interactive demo will need to receive the same treatment. The latter option is to reverse what I was doing and make the actual UI pull from the website directory, which, providing it is light enough, can be pulled onto the device with the rest of the folders. (I was going to exclude the website directory, but now I think I'll forget it).