media queries
The demo has an iframe.
Its container and the iframe both have classes pointing to media queries.
Furthermore, the content of the iframe (_frame pages) have a seperate media query.
This is because the body of the frame pages is having to change. And scrolling/overflow has to be different on the iframe page and the content page, so seperate media queriy stylesheets are being used.